added MicrosoftLearnFilterStore to encapsulate filtering logic; moved the same from MicrosoftLearnStore
