+++ client/src/Features/MicrosoftLearn/MicrosoftLearnSearch.tsx
@@ -9,6 +9,7 @@ import { useObserver } from 'mobx-react-lite';
 export type MicrosoftLearnSearchStyles = SimpleComponentStyles<'root' | 'label' | 'searchBox'>;
 
 const MicrosoftLearnSearchInner = ({ styles }: IStylesOnly<MicrosoftLearnSearchStyles>): JSX.Element => {
+  const learnFilterStore = useStore('microsoftLearnFilterStore');
   const learnStore = useStore('microsoftLearnStore');
 
   const classes = themedClassNames(styles);
@@ -17,10 +18,10 @@ const MicrosoftLearnSearchInner = ({ styles }: IStylesOnly<MicrosoftLearnSearchS
       <div className={classes.root}>
         <Label className={classes.label}>Search</Label>
         <SearchBox
-          onChange={(_e, newValue) => learnStore.updateSearchTerm(newValue || '')}
+          onChange={(_e, newValue) => learnFilterStore.updateSearchTerm(newValue || '')}
           className={classes.searchBox}
           disabled={!!learnStore.isLoadingCatalog}
-          value={learnStore.searchTerm}
+          value={learnFilterStore.searchTerm}
         />
       </div>
     );
