+++ client/src/Stores/Root.store.ts
@@ -4,6 +4,7 @@ import { UsersStore } from './Users.store';
 import { AssignmentLinksStore } from './AssignmentLinks.store';
 import { MicrosoftLearnStore } from './MicrosoftLearn.store';
 import { PlatformStore } from './Platform.store';
+import { MicrosoftLearnFilterStore } from './MicrosoftLearnFilter.store';
 
 export class RootStore {
   assignmentStore = new AssignmentStore();
@@ -11,6 +12,7 @@ export class RootStore {
   assignmentLinksStore = new AssignmentLinksStore();
   microsoftLearnStore = new MicrosoftLearnStore();
   platformStore = new PlatformStore();
+  microsoftLearnFilterStore = new MicrosoftLearnFilterStore();
 
   constructor() {
     const childStores: ChildStore[] = Object.keys(this)
